Push Server support

+++ public/install.php
@@ -1,5 +1,4 @@
 <?php
-
 function pl($v, $a)
 {
     foreach ($a as $value) {
@@ -9,19 +8,7 @@ function pl($v, $a)
     return true;
 }
 
-function json($data)
-{
-    $j = json_encode($data, JSON_UNESCAPED_UNICODE | JSON_UNESCAPED_SLASHES);
-    if (debug) {
-        try {
-            $dat = $j . PHP_EOL;
-            $fp = fopen(dirname(__FILE__) . "/../meta/log.txt", 'a');
-            fwrite($fp, $dat);
-        } catch (Error $e) {
-        }
-    }
-    return $j;
-}
+require_once "./.private/scripts/json.php";
 
 if (pl($_POST, [
     "db", "dbuser", "dbdomain", "dbpass",
@@ -30,9 +17,11 @@ if (pl($_POST, [
 
     class DeleteOnExit
     {
+        static $done = false;
+
         function __destruct()
         {
-            unlink(__FILE__);
+            if (self::$done) unlink(__FILE__);
         }
     }
 
@@ -65,33 +54,42 @@ class DBData
     $db = DB::get();
 
     $db->exec("
-    CREATE TABLE IF NOT EXISTS `roll-in` (
-        `id` INT(11) NOT NULL AUTO_INCREMENT,
-        `token` CHAR(45) NOT NULL COLLATE 'utf8_unicode_ci',
-        `time` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
-        `mono_token` TINYTEXT NOT NULL COLLATE 'utf8_unicode_ci',
-        `proof` TINYTEXT NOT NULL COLLATE 'utf8_unicode_ci',
-        `mono` TINYTEXT NOT NULL COLLATE 'utf8_unicode_ci',
-        PRIMARY KEY (`id`),
-        UNIQUE INDEX `token` (`token`)
-    )
-    COLLATE='utf8_unicode_ci'
-    ENGINE=MyISAM
-    AUTO_INCREMENT=152
-    ;
-    CREATE TABLE IF NOT EXISTS `tokens` (
-        `id` INT(11) NOT NULL AUTO_INCREMENT,
-        `token` VARCHAR(50) NOT NULL COLLATE 'utf8_unicode_ci',
-        `time` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
-        `roll` TINYTEXT NULL COLLATE 'utf8_unicode_ci',
-        `mono` TINYTEXT NOT NULL COLLATE 'utf8_unicode_ci',
-        PRIMARY KEY (`id`),
-        UNIQUE INDEX `token` (`token`)
-    )
-    COLLATE='utf8_unicode_ci'
-    ENGINE=MyISAM
-    AUTO_INCREMENT=55
-    ;
+CREATE TABLE IF NOT EXISTS `roll-in` (
+  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `token` char(45) COLLATE utf8_unicode_ci NOT NULL,
+  `time` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
+  `mono_token` tinytext COLLATE utf8_unicode_ci NOT NULL,
+  `proof` tinytext COLLATE utf8_unicode_ci NOT NULL,
+  `mono` tinytext COLLATE utf8_unicode_ci NOT NULL,
+  PRIMARY KEY (`id`),
+  UNIQUE KEY `token` (`token`)
+) ENGINE=MyISAM AUTO_INCREMENT=853 DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
+
+CREATE TABLE IF NOT EXISTS `subscriptions` (
+  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `type` char(255) COLLATE utf8_unicode_ci NOT NULL,
+  `identificator` char(255) COLLATE utf8_unicode_ci NOT NULL,
+  `token` char(255) COLLATE utf8_unicode_ci NOT NULL,
+  `date` int(11) NOT NULL DEFAULT '0',
+  `state` int(11) NOT NULL DEFAULT '1',
+  `endpoint` text COLLATE utf8_unicode_ci NOT NULL,
+  `key` text COLLATE utf8_unicode_ci NOT NULL,
+  `auth` text COLLATE utf8_unicode_ci NOT NULL,
+  `encoding` text COLLATE utf8_unicode_ci NOT NULL,
+  `expires` int(11) NOT NULL DEFAULT '0',
+  PRIMARY KEY (`id`)
+) ENGINE=MyISAM AUTO_INCREMENT=241 DEFAULT CHARSET=cp1251 COLLATE=utf8_unicode_ci;
+
+CREATE TABLE IF NOT EXISTS `tokens` (
+  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `token` varchar(50) COLLATE utf8_unicode_ci NOT NULL,
+  `time` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
+  `roll` tinytext COLLATE utf8_unicode_ci,
+  `mono` tinytext COLLATE utf8_unicode_ci NOT NULL,
+  `mono_id` text COLLATE utf8_unicode_ci,
+  PRIMARY KEY (`id`),
+  UNIQUE KEY `token` (`token`)
+) ENGINE=MyISAM AUTO_INCREMENT=491 DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
     ");
 
     file_put_contents(__DIR__ . "/.private/meta/priv.key", $_POST["privatekey"]);
@@ -114,28 +112,123 @@ class DBData
     $header_bl = array_filter(explode("|", $_POST["headerbl"]));
     if (count($header_bl) === 0) $header_bl = ["SSL", "GeoIp-Country-Code", "Host"];
 
-    $public_push_cert = (empty($_POST["publicpushcert"]) ? false : $_POST["publicpushcert"]);
-    $private_push_cert = (empty($_POST["privatepushcert"]) ? false : $_POST["privatepushcert"]);
+    $install_log = [];
+    $composer_install_output = [];
+    $composer_require_output = [];
+    $public_push_cert = false;
+    $private_push_cert = false;
+    $pushproof = false;
+
+    if (!($_POST["nopushserver"] === "1")) {
+        if ($_POST["composerdeps"] === "1") {
+            $DIR = __DIR__;
+            $php_executable = (empty($_POST["phpexecutable"]) ? "php" : $_POST["phpexecutable"]);
+
+            if ($_POST["dlcomposer"] === "1") {
+                file_put_contents("../composer-setup.php", file_get_contents("https://getcomposer.org/installer"));
+                exec("export HOME=$DIR/.. && cd .. && $php_executable composer-setup.php 2>&1", $composer_install_output);
+            }
+
+            exec("export HOME=$DIR/.. && cd .. && $php_executable composer.phar install 2>&1", $composer_require_output);
+
+            $install_log = array_merge($install_log, $composer_install_output, $composer_require_output);
+        }
+
+        try {
+            require __DIR__ . '/../vendor/autoload.php';
+            if (!class_exists("Minishlink\WebPush\WebPush")) throw new Exception("No WebPush library");
+            $push_works = true;
+        } catch (Exception $e) {
+            echo "<pre>";
+            echo ("Web Push libraries are not installed. \nDisable Push Server at setup page if you are not going to support this feature.\n\n");
+            foreach ($install_log as $value) {
+                print($value . "\n");
+            }
+            die("</pre>");
+        }
+
+        $pushproof = (empty($_POST["webhooksecret"]) ? bin2hex(random_bytes(10)) : $_POST["webhooksecret"]);
+
+        $public_push_cert = (empty($_POST["publicpushcert"]) ? false : $_POST["publicpushcert"]);
+        $private_push_cert = (empty($_POST["privatepushcert"]) ? false : $_POST["privatepushcert"]);
+
+        if (!$public_push_cert && !$private_push_cert) {
+            $key_gen = Minishlink\WebPush\VAPID::createVapidKeys();
+
+            $public_push_cert = $key_gen["publicKey"];
+            $private_push_cert = $key_gen["privateKey"];
+        }
+    }
+
+    $settings = [
+        "path" => $path,
+        "api_domain" => $api_domain,
+        "headers_blacklist" => $header_bl,
+        "push" => [
+            "public" => $public_push_cert,
+            "private" => $private_push_cert,
+            "proof" => $pushproof,
+        ]
+    ];
 
     file_put_contents(
         __DIR__ . "/.private/meta/settings.json",
-        json(["path" => $path, "api_domain" => $api_domain, "headers_blacklist" => $header_bl, "push" => ["public" => $public_push_cert, "private" => $private_push_cert]])
+        json($settings)
     );
-    exit('
+
+
+    if (!($_POST["nopushserver"] === "1") && $_POST["webhook"] === "1" && $push_works) {
+        require_once __DIR__ . "/.private/scripts/index.php";
+
+        $auth = mono_sender(
+            "/personal/corp/webhook",
+            "POST",
+            null,
+            json(
+                [
+                    "webHookUrl" => (isset($_SERVER['HTTPS']) ? "https" : "http") . "://{$_SERVER['SERVER_NAME']}/push/" . Settings::$get->push->proof . "/accept",
+                ]
+            ),
+            [],
+            "",
+            $api_log
+        );
+        $install_log = array_merge($install_log, $api_log);
+
+        mono_sender(
+            "/personal/corp/settings",
+            "GET",
+            null,
+            json([]),
+            [],
+            "",
+            $api_log
+        );
+        $install_log = array_merge($install_log, $api_log);
+    }
+
+    DeleteOnExit::$done = true;
+    header("Content-Type: text/html");
+    echo ('
 <!DOCTYPE html>
 <html lang="en">
 <head>
     <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta name="viewport" id="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <title>Success</title>
 </head>
 <body>
     <h1 style="margin-bottom: 0">Success</h1>
     <p style="margin-top: 0">Installation script saved the configuration <b>and self-destructed</b></p>
-</body>
+    <details><summary><b>Logs</b></summary><pre>');
+    foreach ($install_log as $value) {
+        echo $value . "\n";
+    }
+    echo ('</pre></details></body>
 </html>    
 ');
+    exit();
 }
 
 ?>
@@ -145,9 +238,22 @@ class DBData
 
 <head>
     <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta name="viewport" id="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <title>Setup</title>
+    <style>
+        label:not(.has-checkbox) {
+            display: inline-block;
+            width: 10em;
+            vertical-align: top;
+            margin-top: 10px;
+        }
+
+        input[type="text"],
+        textarea {
+            margin-top: 10px;
+        }
+    </style>
 </head>
 
 <body>
@@ -157,30 +263,117 @@ class DBData
         <fieldset>
             <legend>Database</legend>
 
-            <input type="text" name="dbdomain" placeholder="Domain"><br>
-            <input type="text" name="db" placeholder="DB"><br>
-            <input type="text" name="dbuser" placeholder="User"><br>
-            <input type="text" name="dbpass" placeholder="Password">
+            <label for="dbdomain">
+                Host:
+            </label>
+            <input type="text" name="dbdomain" id="dbdomain" placeholder="db.example.com"><br>
+
+            <label for="db">
+                Name:
+            </label>
+            <input type="text" name="db" id="db" placeholder="mono_database"><br>
+
+            <label for="dbuser">
+                User:
+            </label>
+            <input type="text" name="dbuser" id="dbuser" placeholder="admin"><br>
+
+            <label for="dbpass">
+                Password:
+            </label>
+            <input type="text" name="dbpass" id="dbpass" placeholder="********">
         </fieldset>
         <fieldset>
             <legend>Signing</legend>
 
-            Private key: <br><textarea name="privatekey" rows="10" cols="70" placeholder="-----BEGIN EC PARAMETERS-----"></textarea><br>
-            <input type="text" name="keyid" size="70" placeholder="Key ID">
+            <label for="privatekey">
+                Private key:
+            </label>
+            <textarea name="privatekey" id="privatekey" rows="10" cols="70" placeholder="-----BEGIN EC PARAMETERS-----"></textarea><br>
+            <label for="privatekey">
+                Key ID:
+            </label>
+            <input type="text" name="keyid" id="keyid" size="70" placeholder="~40 characters long hexadecimal number">
         </fieldset>
         <fieldset>
             <legend>Monobank API</legend>
-            <input type="url" name="apidomain" placeholder="https://api.monobank.ua"><br>
-            Blacklisted headers: <br><input type="text" name="headerbl" size="70" placeholder="SSL|GeoIp-Country-Code|Host">
+            <label for="apidomain">
+                API Endpoint:
+            </label>
+            <input type="url" name="apidomain" id="apidomain" placeholder="https://api.monobank.ua"><br>
+            <label for="headerbl">
+                Blacklisted headers:
+            </label>
+            <input type="text" name="headerbl" id="headerbl" size="70" placeholder="SSL|GeoIp-Country-Code|Host">
         </fieldset>
         <fieldset>
-            <legend>Push API</legend>
-            <pre>
+            <legend>Push Server</legend>
+            <p>
+                <label for="nopushserver" class="has-checkbox">
+                    <input type="checkbox" name="nopushserver" id="nopushserver" value="1">
+                    Disable Push Server and skip this block
+                </label>
+                <details>
+                    <summary><b>Composer will be used to install Web Push libraries. Click to change that.</b></summary>
+                    <label for="composerdeps" class="has-checkbox">
+                        <input type="checkbox" name="composerdeps" id="composerdeps" value="1" checked>
+                        Install Web Push libraries via Composer automatically
+                    </label><br>
+                    <label for="webhook" class="has-checkbox">
+                        <input type="checkbox" name="webhook" id="webhook" value="1" checked>
+                        Set webhook in Monobank API
+                    </label><br>
+                    <label for="dlcomposer" class="has-checkbox">
+                        <input type="checkbox" name="dlcomposer" id="dlcomposer" value="1" checked>
+                        Download Composer on the fly
+                    </label><br>
+                    <label for="phpexecutable">
+                        PHP Executable Path:
+                    </label>
+                    <input type="text" name="phpexecutable" id="phpexecutable" placeholder="php">
+                    <p>If you unchecked the first option, but still want Push Server to work, before proceeding use following command:</p>
+                    <pre>
+    $ composer install
+</pre>
+                    <p>to initiate the project from composer.lock, or install packages manually:</p>
+                    <pre>
+    $ composer require minishlink/web-push 5.2.4
+</pre>
+                    <p>
+                        If you can't use Composer,
+                        download the package from <a href="https://github.com/Sominemo/Mono-Corp-API-Proxy-PHP/releases" target="_blank">GitHub Releases</a>
+                        or <a href="https://php-download.com/" target="_blank">PHP Download</a>
+                        and unpack it to the parent directory of <b>public</b>, so <b>public</b> and <b>vendor</b> folders will be near.
+                    </p>
+                </details>
+            </p>
+            Your Push API keys will be generated automatically
+            <details>
+                <summary><b>Use my own</b></summary>
+                <p>
+                    You can generate your VAPID keys <a href="https://tools.reactpwa.com/vapid" target="_blank">online</a> or by these commands:
+                </p>
+                <pre>
     $ npm install -g web-push
     $ web-push generate-vapid-keys
-            </pre>
-            <input type="text" name="publicpushcert" placeholder="Public Push Cert" size="90"><br>
-            <input type="text" name="privatepushcert" placeholder="Private Push Cert" size="90"><br>
+</pre>
+
+                <label for="publicpushcert">
+                    Public Push Key:
+                </label>
+                <input type="text" name="publicpushcert" id="publicpushcert" placeholder="~87 characters" size="90"><br>
+                <label for="publicpushcert">
+                    Private Push Key:
+                </label>
+                <input type="text" name="privatepushcert" id="privatepushcert" placeholder="~43 characters" size="45"><br>
+            </details>
+            <p>
+                Password for monobank to prevent strangers from sending fake events. If not filled, generates automatically.<br>
+                <label for="webhooksecret">
+                    Webhook secret:
+                </label>
+                <input type="text" name="webhooksecret" id="webhooksecret" placeholder="Generate automatically" size="45"><br>
+            </p>
         </fieldset>
         <fieldset>
             <legend>Write & Set Up</legend>
